Showing multiple images
Now when using "show image", rainbowstream will attempt to download and
save every images in the tweet in a temporary directory, then spawn an
instance of "xv" to show them, and finally destroy them once xv is
closed.

Note that all of this occurs in a thread, thus not pausing the stream.